ct6k: add support for ncurses to emulator
The emulator is now a semi-functional full-screen control panel that
replicates the actual Comp-o-Tron 6000 front panel, as much as is
possible in text mode.

The emulator has support for single step and free run modes at multiple
speeds, as well both traditional blinky lights mode as well as a more
modern 1/0 display. Many features remain unimplemented, but it's a
start.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>