ct6k ui: implment memory display and change functions
Add new functionality to change and display memory contents. Change the
API to use a vector so the user can input multiple words in one go -
useful for manually inputting a boot loader.

Modify API for memory dumping so that the address input call is separate
from the display call, eliminating the need for a dependency on the CPU
object.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>