ct6k ui: add support for disassembvly dump
Similar to memory dump, but this takes a vector of string (for the
instructions) and a vector of uint32 for the addresses. Yes, I could
make a struct of those and then make a single vector of the struct. Yes,
I probably should. But this works, so call that a TODO for future
refactor.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>