ct6k asm: fix issues with symbols
Properly extract symbols from the line instead of blanks (would only
allow a single symbol per program). Begin extraction with the first
character of the symbol, not the $. When replacing symbols with 0, don't
add a newline, this breaks MOV instructions using the symbol address as
a direct value.

Finally, clean up the error messages to make them clearer and properly
indicate the failing source line.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>