ct6k: add peripherals
Add peripheral class and definitions in hw.h to support devices.
Currently only the Print-o-Tron XL is supported but the framework is in
place to add the Card-o-Tron 3CS later.

Introduces "Periph" as an abstract class that can be instantiated by
concrete classes representing the devices. This allows the CPU class to
use polymorphism to make make common calls to the devices.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>