ct6k cpu: add support for peripherals
All Comp-o-Tron peripherals are memory mapped and the io region location
is determined at boot time. Add support to the CPU class for registering
and removing peripherals (hotplug not supported!) and assigning them
memory regions.

Modify the memory read and write functions to pass requests to
peripherals as needed, and to allow reading of the peripheral mapping
table.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>