ct6k QT: Add ugly GUI
This is possibly the ugliest GUI created by anyone since 1987, but it
works. Printer output is not yet supported, but all other features are
present and accounted for, including examining and modifying registers
and memory, and disassembly support.

QT defaults to naming C++ headers as foo.h, but these are all named .hpp
in keeping with the coding standard of the rest of the project.

At full speed, this executes about 600% faster than the text UI.

Known issues:
 - font loading does not work correctly
 - ugly

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>