ct6k gui: add printer support
Add support for the Print-o-Tron XL full-width matrix imager.
The printer is emulated in a window showing pink and white bars like
standard Comp-o-Tron pinkbar paper. By default the window is hidden,
and is shown via a menu option. Any available output is sent to the
printer at the same time that the control panel is updated.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>