ct6k gui: set backgroun color correctly
Setting the background color on the main window has some weird
side-effects, like all of the menus inheriting that color as well. The
correct way to do it is to have the control panel create a colored frame
widget that is exactly the size of the main window, and set that to the
desired color.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>