ct6k gui: remove check from printer menu
Remove the checkbox from the menu item that displays the Print-o-Tron XL
window. There's no need to close the window from this menu - the window
has an X on the titlebar that works fine and is much more natural. So we
lose the checkbox and just uncondionally call show() when the menu item
is selected.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>