ct6k gui: add card reader support
Add gui support for the emulation of the Card-o-Tron 3CS Punch Card
Scanner and Puncher! This big patch adds a new window, triggered by a
menu option, and uses exciting new *red* blinky lights. Like the printer
class, the card reader class is istantiated along with the CPU and
status updates are signaled by the CPU spinner.

Unlike the printer emulation, we need two-way communication between the
Card-o-Tron window and the actual Card-o-Tron class, because the window
is the mechanism by which the user opens files for I/O.

This is largely untested. The actual reading and writing has been unit
tested, and the menu/window/buttons have been tested. The missing piece
is writing assembly code to actually exercise the device.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>