ct6k punch: add card puncher utility
This simple filter takes text from standard input and spits out
Card-o-Tron records to standard output. Redirect that output into a .ctc
file and you've got a card deck that can be fed to the emulator.

Do not use this for binary or object code! The standard library will
mangle it!

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>