ct6k: add makefile for Windows
Add a makefile for nmake to build the assembler and card punch
command-line tools. This must be executed from within the Developer
Command Prompt environment. This hard-codes the Boost install directory;
you will need to change it if you use a different version of Boost or
install Boost to a different directory.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>