ct6k gui: scale down for small screens
On startup, detect the screen size and scale down to fit small (1080p)
screens. Sorry, 720p folks, you still don't get to play.

The control panel object detects the screen size and passes a
SmallScreen flag down to all of its elements, allowing them to scale
down. For the most part, this is propagated down to the Indicators,
which shrink themselves by 25%.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>