ct6k gui: allow register params for debug display
Allow input of a register name to provide the address for the Dump
Memory and Disassemble functions. This is a nice convenience - if R0
contains the address of a buffer, just enter R0 to dump the buffer. Or
easily disassemble from where the program sits by entering IP instead of
converting from blinkenlights to hex in your head.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>