ct6k gui: add scripts to support installer
win-config-installer.sh builds the necessary tree and kicks off the QT
installer tool to make an exe that will install the emulator. It's ugly
and the paths are all hard-coded and bad, but it works.
win-install-clean.sh cleans up all of the files copied and created by
the config script.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>