asm6k: add new segment class
This is the start of a major overhaul to the assembler. In order to
support assembling multiple files into a single program, we need to
track individual pieces of the code, which are called Code Segments.
Yes, this overloads a phrase used in low-level x86 programming, but it
makes sense in this context.

This class just keeps track of a segment of binary code and the file it
came from. One or more segments make up a program, and are linked
together as the final step of the assembly process.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>