asm6k: assemble multiple source files
This major upgrade to the assembler allows the user to stack together
multiple assembler source files into a single program, much like a user
in the 1950s would stack together several decks of cards to make a
single program.

To support that goal, we add support for multiple segments of binary
code, referred to as Code Segments. Each source file assembles to one or
more Code Segments, which are then linked together at the end.

To support more flexible and mature programs, this change also
introduces new assembler directives:
  .VALUE defines a constant much like a C #define
  .ADDR sets the destination address of the ensuing binary code or data
  .TXT[M|L|N] to simplify the inclusion of text

Additionally, the assembler now supports output of binary code to a
stack of Card-o-Tron cards that can be read by the emulator. This format
allows non-contiguous code segments and will give the flexibility to
write complex libraries and program loaders.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>