asm6k: produce listing file
Add the -l option to produce a listing of the assembled binary file. The
listing file is a text file with  the same name as the output file, with
.listing appended. This file contains the raw hex data of the program,
along with its associated assembly code. This is the exact code as run
by the Comp-o-Tron 6000, listed in the exact order at the exact address.

A proper listing file is invaluable for debugging!

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>