ct6k cpu: add BRK instruction
This instruction just indicates that processing should stop - no flags
are affected and the system is able to resume immediately. It's intended
for debugging, but may also be used to give the user a chance to change
cards or tapes before continuing.

Note that this is just the instruction and CPU part - the caller must
check the Broken flag after each instruction.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>