examples: remove assumptions from readcard
Readcard depended on both R11 and R12 being set to Card-o-Tron scanner
registers, and R5 having the card info. This is unnecessary; if we have
one, we can get the others. Quit putting the burden on the caller.

Also added a comment about the device being present and ready with card
data.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>