ct6k: print everything
If the CPU hits a state makes it stop cycling (like HALT), it would
ignore any pending printer output until it cycled again. In the case of
HALT, this is never, so the last output line might just get eaten.

To fix, this check for multiple lines of output each time. In general,
there will be only one, but add a while loop to make sure we get all of
the output. Belt and suspenders.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>