ct6k print-o-tron: hang on to multiple output lines
If the GUI doesn't read our output line quickly enough, we might lose
it. This is particularly true if we have a line release immediately
followed by a page release.

To fix this, use a vector of std::string to store output text. Normally,
this would hold only one line, but this allows us to stack up an
arbitrary number of text lines without worrying about losing anything
due to a slow GUI.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>