asm6k: output ROM images for the emulator
Add an option for the assembler to output a binary file as a C++ header
for inclusion into the emulator. This allows us to create a boot ROM
image in order to support the Comp-o-Tron 6000 Self-Rising Flow-er.

The output file is first output as a binary file so that all of the
necessary checks are done. If it's good, then the assembler will
re-output the binary image as a header file, with a #define for the
start address.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>