ct6k: build a boot ROM
Add a makefile target to make our boot ROM header file, and the source
to go along with it. This ROM loads at an area well above main memory.
Once executed, it reads a stack of cards (type Code) into main memory,
then jumps to the address found on the first card.

To execute, the user loads a stack of cards (containing, hopefully, an
executable image) into the Card-o-Tron Scanner, then sets the IP
register to the ROM address (0xFFC00000) and presses RUN. If a printer
is available, the ROM will print information messages.

Just before the final step of jumping to the loaded program, the ROM
executes a BRK instruction, giving the user the opportunity to load a
new card deck before continuing.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>