ct6k: add ROM image
The ROM image is owned by the control panel and is loaded at the time
the CPU is instantiated. No checking is done, and the image is not
executed until the user starts it manually.

Currently, this header file is created with the makefile from the
command-line tools; this must be built before the emulator is built.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>