ct6k: get rid of Boost dependencies
Part one of simplifying the build process. I was only using Boost for a
very few simple things that could be easily coded using the C++ STL.

The output functions are bit clunky but are easily moved into a helper
function so that the code flow isn't overly complicated.

The tokenizer was vastly underused anyway in this setting, so it was
replaced with a simple text extraction function.

Signed-off-by: Mitch Williams <cluemerchant@gmail.com>